--- langchain-local-knowledgebase/.env.example
@@ -1,3 +1,0 @@
-PORT = 3200 # Port to run the server on
-OPENAI_API_KEY="sk-" # OpenAI API key
-SECRET = "secret" # Secret for encoding
