+++ langchain-local-knowledgebase/README.md
@@ -1,66 +1,26 @@
-# Knowledge Base Demo | Voiceflow
-# Open AI GPT, Langchain, HNSWLib and Cheerio
-
-This code utilizes Open AI GPT, Langchain, HNSWLib and Cheerio to fetch web content from URLs, create embeddings/vectors and save them in a local database. The knowledge base can then be used with GPT to answer questions.
-
-## Quickstart Video
-[![Watch the video](https://s3.amazonaws.com/com.voiceflow.studio/share/CleanShot-2023-03-17-at-14.03.29/CleanShot-2023-03-17-at-14.03.29.png)](https://www.loom.com/share/a4fccc7aac7d48548006570f6ac98576)
+# Langchain API for Voiceflow
 
 
-## Node.js
-If you are running this on Node.js 16, either:
+## Getting Started
 
-run your application with NODE_OPTIONS='--experimental-fetch' node ..., or
-install node-fetch and follow the instructions <a href="https://github.com/node-fetch/node-fetch#providing-global-access" target="_blank" rel="noopener noreferrer">here</a>
+### Cloning the repository
 
-If you are running this on Node.js 18 or 19, you do not need to do anything.
-
-
-## Setup
-Setup is simple, just run:
-
-```
-npm i
-npm start
+```bash
+git clone https://github.com/voiceflow-gallagan/vf-langchain-api.git && cd vf-langchain-api
 ```
 
-This app requires an `.env` file with `PORT`, `OPENAI_API_KEY` and `SECRET` (for API key encoding).
-You can rename the .env.example file to .env and fill in the values.
+### Create and edit the .env file
 
-## Endpoints
-
-There are 2 endpoints available:
-
-### `/parser`
-
-This endpoint allows a POST request with the URL link of the page to parse.
-You need to pass the Voiceflow Assistant API Key in the body.
-
-```json
-{
-	"url":"https://www.voiceflow.com/blog/prompt-chaining-conversational-ai",
-	"apikey":"VF.DM.XXX"
-}
+```bash
+cp .env.example .env
 ```
 
-### `/question`
+In the .env file, you will need to add your **OPENAI_API_KEY**.
+You can change the **SECRET** and the **PORT** values if you want to.
 
-This endpoint allows a POST request with the question.
-You need to pass the Voiceflow Assistant API Key in the body.
+### Installation (1st time)
 
-```json
-{
-	"question":"What is prompt chaining?",
-	"apikey":"VF.DM.XXX"
-}
-```
+  ```bash
+  npm run build
+  ```
 
-## Using ngrok
-
-To allow access to the app externally using the port set in the `.env` file, you can use ngrok. Follow the steps below:
-
-1. Install ngrok: https://ngrok.com/download
-2. Run `ngrok http <port>` in your terminal (replace `<port>` with the port set in your `.env` file)
-3. Copy the ngrok URL generated by the command and use it in your Voiceflow Assistant API step.
-
-This can be handy if you want to quickly test this in an API step within your Voiceflow Assistant.
