+++ langchain-local-knowledgebase/app/.env.example
@@ -1,0 +1,5 @@
+PORT = 3000 # Port to run the server on
+OPENAI_API_KEY="sk-" # OpenAI API key
+SECRET = "secret" # Secret for encoding
+UNSTRUCTURED_URL = "http://localhost:8000"
+REDIS_URL = "redis://localhost:6379"
