+++ langchain-local-knowledgebase/app/.gitignore
@@ -1,8 +1,7 @@
-.chroma
-chroma
-.env
-volumes
+# NiKo
 backup
+tmp
+collections
 
 # Logs
 logs
@@ -11,6 +10,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
+
 
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
@@ -108,3 +108,4 @@ dist
 
 # TernJS port file
 .tern-port
+package-lock.json
