+++ langchain-local-knowledgebase/app/docker-compose.yml
@@ -1,0 +1,26 @@
+version: '3.9'
+
+services:
+
+  unstructured:
+    image: 'quay.io/unstructured-io/unstructured-api:latest'
+    container_name: unstructured
+    hostname: unstructured
+    ports:
+      - '8000:8000'
+    networks:
+      - net
+
+  redis:
+    image: 'redis:6.2-alpine'
+    container_name: redis
+    command: ["redis-server", "--bind", "redis", "--port", "6379"]
+    hostname: redis
+    ports:
+      - '6379:6379'
+    networks:
+      - net
+
+networks:
+  net:
+    driver: bridge
